added JDBC dao and created some tests

+++ src/main/java/com/veryan/FlooringMastery/controller/FlooringController.java
@@ -10,26 +10,32 @@ import com.veryan.FlooringMastery.veiw.Veiw;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
+import java.time.LocalDate;
 import java.util.List;
 
+
+/**
+ * the controller for the application
+ * communicates with vew and service to run the program
+ */
 @Controller
 public class FlooringController {
     private final Service service;
-    private final Veiw consoleVeiw;
+    private final Veiw veiw;
     private boolean running = true;
     private String methodRepeat = "";
 
     @Autowired
-    public FlooringController(Service service, Veiw consoleVeiw) {
+    public FlooringController(Service service, Veiw veiw) {
         this.service = service;
-        this.consoleVeiw = consoleVeiw;
+        this.veiw = veiw;
     }
 
     public void run() {
         while (running) {
             String action;
             if (methodRepeat.isEmpty()) {
-                action = consoleVeiw.menu().toLowerCase();
+                action = veiw.menu().toLowerCase();
             } else {
                 action = methodRepeat;
                 methodRepeat = "";
@@ -49,7 +55,7 @@ public class FlooringController {
                     remove();
                     break;
                 case "export":
-                    consoleVeiw.userMessage("exported data");
+                    veiw.userMessage("exported data");
                     service.exportData();
                     break;
                 case "quit":
@@ -61,58 +67,91 @@ public class FlooringController {
     }
 
     private void display(){
-        List<Order> orders = service.getOrders();
-        consoleVeiw.displayOrders(orders);
+        LocalDate date = veiw.askDisplayOrders();
+
+        if(date == null){
+            veiw.userMessage("invalid date format");
+            return;
+        }
+
+        try {
+            List<Order> orders = service.getOrders(date);
+            veiw.displayOrders(orders);
+        } catch (NoSuchOrder e) {
+            veiw.userMessage("no orders");
+        }
     }
 
     private void add(){
-        Order order = consoleVeiw.makeOrder();
+        Order order = veiw.makeOrder(service.getProducts());
         try {
-            service.addOrder(order);
+            order = service.initliseOrder(order);
+            veiw.displayOrder(order);
+            boolean answer = veiw.userQuestion("add order? y/n");
+            if(answer) {
+                service.addOrder(order);
+            }
         } catch (InvalidInput e) {
-            consoleVeiw.userMessage("Invalid input, tax or product is incorrect");
+            veiw.userMessage("Invalid input, "+e.getMessage());
             this.methodRepeat = "add";
         } catch (OrderAlreadyExists e) {
-            boolean answer = consoleVeiw.userQuestion("Order already exists, return to menu? y/n");
+            boolean answer = veiw.userQuestion("Order already exists, return to menu? y/n");
             if(!answer){this.methodRepeat="add";}
         }
     }
     private void edit(){
-        OrderID orderID = consoleVeiw.findOrder();
+        OrderID orderID = veiw.findOrder();
+        if(orderID.date() == null || orderID.orderNumber() < 0){
+            veiw.userMessage("Invalid input");
+            this.methodRepeat="edit";
+            return;
+        }
+
         Order order;
         try {
             order = service.getOrder(orderID.date(), orderID.orderNumber());
         } catch (NoSuchOrder e) {
-            boolean answer = consoleVeiw.userQuestion("Cant find Order, return to menu? y/n");
+            boolean answer = veiw.userQuestion("Cant find Order, return to menu? y/n");
             if(!answer){this.methodRepeat="edit";}
             return;
         }
 
-        consoleVeiw.userMessage("Old Order");
-        consoleVeiw.displayOrder(order);
-        consoleVeiw.userMessage("input the changed data (cant change date)");
-        Order newOrder = consoleVeiw.makeOrder();
+        veiw.userMessage("Old Order");
+        veiw.displayOrder(order);
+        veiw.userMessage("input the changed data (cant change date)");
+        Order newOrder = veiw.makeOrder(service.getProducts(), order.date);
+        newOrder.orderNumber = order.orderNumber;
 
         try {
-            service.replaceOrder(order, newOrder);
+            newOrder = service.initliseOrder(newOrder);
+            veiw.displayOrder(newOrder);
+            boolean answer = veiw.userQuestion("replace order? y/n");
+            if(answer) {
+                service.replaceOrder(order, newOrder);
+            }
         } catch (NoSuchOrder ignored) {
             //can't be reached
         } catch (InvalidInput e) {
-            boolean answer = consoleVeiw.userQuestion("Invalid input for new order, tax or product is incorrect, return to menu? y/n");
+            boolean answer = veiw.userQuestion("Invalid Input, "+e.getMessage()+" return to menu? y/n");
             if(!answer){this.methodRepeat="edit";}
         }
     }
 
     private void remove(){
-        OrderID orderID = consoleVeiw.findOrder();
+        OrderID orderID = veiw.findOrder();
         Order order;
         try {
             order = service.getOrder(orderID.date(), orderID.orderNumber());
+            veiw.displayOrder(order);
         } catch (NoSuchOrder e) {
-            consoleVeiw.userMessage("no such order");
+            veiw.userMessage("no such order");
             return;
         }
 
-        service.deleteOrder(order);
+        boolean answer = veiw.userQuestion("delete order? y/n");
+
+        if(answer) {
+            service.deleteOrder(order);
+        }
     }
 }
